Addresses review feedback
