+++ src/app/proposals/draft/components/BasicProposalForm.tsx
@@ -219,7 +219,6 @@ const BasicProposalForm = () => {
         }
       });
     } catch (e) {
-      console.error(e);
       toast.error("Error simulating transactions");
     } finally {
       setFormDirty(false);
