+++ src/app/proposals/draft/components/stages/DraftForm/DraftFormClient.tsx
@@ -157,12 +157,18 @@ const DraftFormClient = ({
                   control={control}
                   label="Proposal type"
                   required={true}
-                  options={validProposalTypes.map((typeConfig) => {
-                    return {
-                      label: typeConfig.name,
-                      value: typeConfig.proposal_type_id,
-                    };
-                  })}
+                  //   options={validProposalTypes.map((typeConfig) => {
+                  //     return {
+                  //       label: typeConfig.name,
+                  //       value: typeConfig.proposal_type_id,
+                  //     };
+                  //   })}
+                  options={[
+                    {
+                      label: "default",
+                      value: "0",
+                    },
+                  ]}
                   name="proposalConfigType"
                   emptyCopy="Default"
                 />
